// frontend/src/lib/api/generated/model/manageUserDTO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ระบบรายงานค่าสาธารณูปโภค กรมประมง API
 * OpenAPI spec version: 2.0.0
 */

export interface ManageUserDTO {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  banned: boolean;
  createdAt: string;
  departmentId?: string;
  departmentName?: string;
  email: string;
  emailVerified: boolean;
  hasPassword: boolean;
  id: string;
  name: string;
  phone?: string;
  role: string;
}